--- Develop/server.js
@@ -1,21 +1,0 @@
-const express = require('express');
-const inquirer = require('inquirer');
-const mysql = require('mysql2');
-
-const PORT = process.env.PORT || 3001;
-const app = express();
-
-// middleware
-app.use(express.urlencoded({ extended: false }));
-app.use(express.json());
-
-// connect to database
-const db = mysql.createConnection(
-  {
-    host: 'localhost',
-    user: 'root',
-    password: 'password',
-    database: 'employees_db'
-  },
-  console.log('Connected to employees_db.')
-);
